Initial implementation startup
Setup primary list view that will show the meal objects for users to choose from.
Setup basics for URLSession so we can begin making API calls
Initial exploration of the mealdb API.
Thinking through how I want to deal with integrating the details of a Meal, as the layout of several of the properties in this API are not ideal.

Note: Tested AsyncImage and found that caching is not built in yet, pulling in Kingfisher as it does superior job for caching.