Meal Details addition & UI Updates
Added in basics for the Meal Details request. Currently keeping Details and Meal separated, but no significant reason for this beyond keeping the loading of these instances detached. Maybe some value in simplicity. Obvious answer is to back data structure with Core Data/Swift Data and merge them together.

Tweaked Meal UI to look slightly better, but still has some weirdness in its sizing/spacing that is less than ideal. Need to decide on variability vs. uniformity when it comes to sizing/spacing.

MealDetail's API data structure is not ideal, masking this with an inter-op API structure that can be converted to something more pleasant for the application. Need to consider collapsing ingredient and measurements into a unified object to resolve some UI layout pains with Strings and ForEach identity requirements.