Code Review
Going over the basics, noticed that when appending the final component of the path on the base URL for a Request, I had an inconsistency. It seems appending(path:)'s implementation will automatically insert the required /, or if it is provided it will avoid that behavior.

Adding a / for consistency to getMealDetails, as providing the proper / feels safer.